refactor(basic-select): изменить структуру ФС
