--- src/components/BasicSelect/index.ts
@@ -1,1 +1,0 @@
-export { BasicSelect } from './BasicSelect';
